support PDF -> PNG conversion internally via XPDF
Converting via an external tool means that we have to install more
tools which can be difficult on machines where one does not have
admin access. Additionally, converting with an external tool is
slow.

XPDF seems fine to use for now, and it's smaller than PDFium.
If we ever switch, https://github.com/reMarkable/pdfium-prebuilt/
has prebuilt binaries that could be useful.

One reason would be license (GPL vs BSD), but GPL isn't really an
issue for this project.